search for users works
maybe refine the looks of a few things as profile pictures are not always working, the add button changes position depending on the length of the name ect.

Currently if you click the user it will show a toast, maybe if we set it so that if you press the user, it takes you to their profile page? just a thought.